tools/routa-fitness/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "routa-fitness"
version = "0.1.0"
description = "Evolutionary architecture fitness engine for Routa"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0,<7",
]

[project.optional-dependencies]
graph = [
    "code-review-graph>=1.8,<2",
]
mcp = [
    "fastmcp>=0.1.0,<2",
]
dev = [
    "pytest>=8.0,<9",
    "ruff>=0.3.0,<1",
]

[project.scripts]
routa-fitness = "routa_fitness.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["routa_fitness"]

[tool.ruff]
line-length = 100
target-version = "py310"
